Fixing loginmodal.
