+++ frontend/src/components/modals/bet/BetUtils.jsx
@@ -1,24 +1,13 @@
+// import API_URL from your config
 import { API_URL } from '../../../config';
-import { useEffect, useState } from 'react';
 
-export const submitBet = () => {
+export const submitBet = (betData, token, onSuccess, onError) => {
 
-    const [token, setToken] = useState(null);
-
-    useEffect(() => {
-        setToken(localStorage.getItem('token'));
-    }, []);
 
     if (!token) {
         alert('Please log in to place a bet.');
         return;
     }
-
-    const betData = {
-        marketId,
-        amount: betAmount,
-        outcome: selectedOutcome,
-    };
 
     console.log('Sending bet data:', betData);
 
@@ -29,17 +18,19 @@ export const submitBet = () => {
             Authorization: `Bearer ${token}`,
         },
         body: JSON.stringify(betData),
-    }).then(response => {
+    })
+    .then(response => {
         if (!response.ok) {
             throw new Error(`HTTP error! Status: ${response.status}`);
         }
         return response.json();
-    }).then(data => {
+    })
+    .then(data => {
         console.log('Success:', data);
-        alert(`Bet placed successfully! Bet ID: ${data.id}`);
-        setShowBetModal(false);
-    }).catch(error => {
+        onSuccess(data);  // Handle success outside this utility function
+    })
+    .catch(error => {
         console.error('Error:', error);
-        alert(`Error placing bet: ${error.message}`);
+        onError(error);  // Handle error outside this utility function
     });
 };
